Refractor entity spawning to allow chaning a pet's entity
As of 1.11 each guardian, horse, zombie, and skeleton type has it's own
unique entity type. Instead of being able to switch the type of the
current entity, we have to respawn these entities when we want to switch
types. The advantage of this approach is it allows us to keep
compatibility with older versions that don't have seperate types.

For now, just the guardian properly switches between elder/normal,
but more will come soon!